FirstResultLinter should request first_result URL
The FirstResultLinter is responsible -- right now -- for taking the URL
returned by the searcher and using Net::HTTP for returning its source.

I may refactor this out depending on the complexity of the implementation
-- the FirstResultLinter should remain an Orchestrator and not include
implementation details itself.

Add WebMock to the Gemfile to mock and set expectations on web requests.